requirements.txt
asgiref==3.11.1
Django==5.2.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
pillow==12.2.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.3
gunicorn==23.0.0
anthropic==0.52.0
